+++ SuperNodes/src/SuperNodesGenerator.cs
@@ -9,211 +9,7 @@ using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.CodeAnalysis.Text;
-
-/// <summary>
-/// Represents an item given to the generator's Execute method. Because of the
-/// way incremental generator pipelines work, this is used to represent the
-/// information that is needed to generate the implementation of a single
-/// SuperNode.
-/// </summary>
-/// <param name="Node">The SuperNode description to be generated.</param>
-/// <param name="PowerUps">A dictionary of PowerUps, keyed by the PowerUp's
-/// fully resolved name (e.g., <c>global::SomeNamespace.SomePowerUp</c>).
-/// </param>
-public record GenerationItem(
-  SuperNode Node,
-  ImmutableDictionary<string, PowerUpDescription> PowerUps
-);
-
-/// <summary>
-/// Description of an attribute applied to a class field or property member.
-/// </summary>
-/// <param name="Name">Name of the attribute.</param>
-/// <param name="Type">Type of the attribute.</param>
-/// <param name="ArgumentExpressions">Argument expressions given to the
-/// attribute.
-/// </param>
-public record AttributeDescription(
-  string Name,
-  string Type,
-  string[] ArgumentExpressions
-);
-
-/// <summary>
-/// Field or property information. Script properties are captured into a static
-/// table for reflection-style lookups. This allows power-ups to be more, uh,
-/// powerful.
-/// </summary>
-/// <param name="Name">Property or field name.</param>
-/// <param name="Type">Property or field type.</param>
-/// <param name="Attributes">Basic attribute information.</param>
-/// <param name="IsField">True if a field, false if a property.</param>
-public record PropOrField(
-  string Name,
-  string Type,
-  IImmutableSet<AttributeDescription> Attributes,
-  bool IsField
-);
-
-/// <summary>
-/// Represents a lifecycle method handler provided by another source generator
-/// or a PowerUp's lifecycle method handler.
-/// </summary>
-public interface IGodotNodeLifecycleHook { }
-
-/// <summary>
-/// A string containing the name of the method that would be generated by
-/// another source generator or provided by a partial implementation.
-/// </summary>
-/// <param name="Name">Expected method name.</param>
-public record LifecycleMethodHook(string Name) : IGodotNodeLifecycleHook;
-
-/// <summary>
-/// Represents a lifecycle method handler provided by an applied PowerUp.
-/// </summary>
-/// <param name="FullName">Fully resolved name of the PowerUp.</param>
-/// <param name="TypeArguments">Type arguments provided to the PowerUp, if the
-/// PowerUp has generic parameters. Otherwise an empty array.</param>
-public record PowerUpHook(string FullName, ImmutableArray<string> TypeArguments)
-  : IGodotNodeLifecycleHook;
-
-/// <summary>
-/// Represents a SuperNode. A SuperNode is a Godot node script class defined by
-/// a game developer. SuperNodes generate static tables containing information
-/// about their fields and properties so that reflection is not needed to
-/// access that information, in addition to calling declared lifecycle method
-/// handlers that may be provided by other source generators or partial class
-/// implementations. SuperNodes also allow PowerUps to be applied to them,
-/// similar to mixins in other languages (and somewhat like templates in C++).
-/// </summary>
-/// <param name="Namespace">Fully qualified namespace containing the node
-/// (without the <c>global::</c> prefix).</param>
-/// <param name="Name">Name of the Godot Node (not fully qualified). Combine
-/// with <paramref name="Namespace" /> to determine the fully resolved name.
-/// </param>
-/// <param name="Location">The location of the class declaration syntax node
-/// that corresponds to the SuperNode.</param>
-/// <param name="BaseClasses">Array of fully qualified base types (base class
-/// and implemented interfaces).</param>
-/// <param name="LifecycleHooks">Lifecycle hooks that need to be invoked from
-/// the SuperNode's generated <c>_Notification(long what)</c> method.
-/// Lifecycle hooks can either be the string names of declared methods to invoke
-/// (as specified in the SuperNode attribute) or a PowerUp to apply to the
-/// SuperNode.</param>
-/// <param name="PowerUpHooksByFullName">Applied PowerUps (as specified in the
-/// SuperNode attribute), keyed by the fully resolved name of the PowerUp.
-/// </param>
-/// <param name="NotificationHandlers">Notification handlers found in the
-/// script class, such as <c>OnEnterTree</c>, <c>OnProcess(double delta)</c>,
-/// <c>OnReady</c>, etc.</param>
-/// <param name="HasPartialNotificationMethod">True if the SuperNode has a stub
-/// for <c>public override partial void _Notification(long what)</c>. Since
-/// SuperNodes cannot work without this stub, we track whether or not it exists
-/// so we can issue a warning at generation time if it is missing.</param>
-/// <param name="HasNotificationMethodHandler">SuperNode scripts can implement
-/// a special notification handler, <c>OnNotification(long what)</c> that will
-/// be called whenever any notification is received. Since this is a special
-/// case, we track it separately.</param>
-/// <param name="PropsAndFields">Information about properties and fields
-/// declared in the SuperNode. We track these fields so that static reflection
-/// tables can be generated at build time, allowing scripts and PowerUps to
-/// introspect their properties and fields (and the attributes applied to them)
-/// without having to use reflection.</param>
-/// <param name="Usings">All of the using imports defined for the SuperNode
-/// script.</param>
-public record SuperNode(
-  string Namespace,
-  string Name,
-  Location Location,
-  ImmutableArray<string> BaseClasses,
-  ImmutableArray<IGodotNodeLifecycleHook> LifecycleHooks,
-  ImmutableDictionary<string, PowerUpHook> PowerUpHooksByFullName,
-  ImmutableArray<string> NotificationHandlers,
-  bool HasPartialNotificationMethod,
-  bool HasNotificationMethodHandler,
-  ImmutableArray<PropOrField> PropsAndFields,
-  IImmutableSet<string> Usings
-) {
-  /// <summary>
-  /// Filename prefix to use when generating the SuperNode's related
-  /// implementation files.
-  /// </summary>
-  public string FilenamePrefix
-    => Namespace is not "" ? $"{Namespace}.{Name}" : Name;
-}
-
-/// <summary>
-/// Represents a PowerUp. A PowerUp is a class that can be applied to a
-/// SuperNode. Unlike default interface implementations, PowerUps can add
-/// additional state to a SuperNode, similar to mixins or traits in other
-/// programming languages. Essentially, anything in a PowerUp's source code
-/// is copied over into the implementation of the SuperNode that it is applied
-/// to.
-/// </summary>
-/// <param name="Namespace">Fully qualified namespace containing the node
-/// (without the <c>global::</c> prefix).</param>
-/// <param name="Name">Name of the PowerUp class (not fully qualified). Combine
-/// with <paramref name="Namespace" /> to determine the fully resolved name.
-/// </param>
-/// <param name="FullName">Fully resolved name of the PowerUp class.</param>
-/// <param name="Location">The location of the class declaration syntax node
-/// that corresponds to the PowerUp.</param>
-/// <param name="BaseClass">Fully resolved base class of the PowerUp class,
-/// if any. The base class is used as a constraint when applying the PowerUp
-/// to a SuperNode. If the SuperNode the PowerUp is being applied to isn't a
-/// subtype of the PowerUp's base class, an error is shown to the developer.
-/// </param>
-/// <param name="TypeParameters">An array of the declared type parameters for
-/// the PowerUp class (in order), if any.</param>
-/// <param name="Interfaces">Array of fully qualified interfaces implemented by
-/// the PowerUp class. When applying a PowerUp to a SuperNode, any interfaces
-/// implemented by that PowerUp are also applied to the SuperNode.</param>
-/// <param name="Source">Source text of the PowerUp. This is re-parsed during
-/// generation and adapted into the code needed to apply the PowerUp to a given
-/// SuperNode.</param>
-/// <param name="PropsAndFields">Information about properties and fields
-/// declared in the PowerUp. We track these fields so that static reflection
-/// tables can be generated at build time, allowing scripts and PowerUps to
-/// introspect their properties and fields (and the attributes applied to them)
-/// without having to use reflection.</param>
-/// <param name="Usings">All of the using imports defined for the PowerUp
-/// script.</param>
-/// <param name="HasOnPowerUpMethod">True if the PowerUp has declared a
-/// lifecycle method handler that should be called when the SuperNode it has
-/// been applied to receives a Godot lifecycle notification.</param>
-public record PowerUpDescription(
-  string Namespace,
-  string Name,
-  string FullName,
-  Location Location,
-  string BaseClass,
-  ImmutableArray<string> TypeParameters,
-  ImmutableArray<string> Interfaces,
-  string Source,
-  IList<PropOrField> PropsAndFields,
-  IImmutableSet<string> Usings,
-  bool HasOnPowerUpMethod
-) {
-  /// <summary>True if the PowerUp has generic parameters.</summary>
-  public bool IsGeneric => TypeParameters.Length > 0;
-}
-
-/// <summary>
-/// Represents a lifecycle method on a node script (i.e., a method that should
-/// be invoked when a Godot node or object receives a notification from the
-/// Godot engine).
-/// </summary>
-/// <param name="GodotNotification">The name of the Godot node or object
-/// notification.</param>
-/// <param name="ReturnType">The return type of the method that handles the
-/// Godot node or object notification.</param>
-/// <param name="Args">Arguments that should be given to the notification
-/// handler.</param>
-public readonly record struct LifecycleMethod(
-  string GodotNotification,
-  string ReturnType,
-  ImmutableArray<string> Args
-);
+using SuperNodes.Models;
 
 [Generator]
 public partial class SuperNodesGenerator
@@ -326,7 +122,7 @@ public partial class SuperNodesGenerator
     .SelectMany(list => list.Attributes)
     .Any(attribute => attribute.Name.ToString() == POWER_UP_ATTRIBUTE_NAME);
 
-  public static PowerUpDescription GetPowerUpSyntaxCandidate(
+  public static PowerUp GetPowerUpSyntaxCandidate(
     GeneratorSyntaxContext context, CancellationToken _
   ) {
     var node = (ClassDeclarationSyntax)context.Node;
@@ -391,7 +187,7 @@ public partial class SuperNodesGenerator
 
     Log.Print("Registering power up " + fullName);
 
-    return new PowerUpDescription(
+    return new PowerUp(
       Namespace: @namespace,
       Name: name,
       FullName: fullName,
@@ -588,7 +384,7 @@ public partial class SuperNodesGenerator
     );
 
     var powerUps = item.PowerUps;
-    var appliedPowerUps = new List<PowerUpDescription>();
+    var appliedPowerUps = new List<PowerUp>();
 
     // See if the node has any power-ups.
     foreach (var lifecycleHook in node.LifecycleHooks) {
@@ -726,7 +522,7 @@ public partial class SuperNodesGenerator
   }
 
   public static string GenerateSuperNodeStatic(
-    GenerationItem item, IList<PowerUpDescription> appliedPowerUps
+    GenerationItem item, IList<PowerUp> appliedPowerUps
   ) {
     var node = item.Node;
 
@@ -866,7 +662,7 @@ public partial class SuperNodesGenerator
   }
 
   public static string GeneratePowerUpImplementation(
-    SuperNode node, PowerUpDescription powerUp
+    SuperNode node, PowerUp powerUp
   ) {
     // Edit the pieces of the user's power-up needed to make it suitable to be
     // a partial class of the specific node script it's applied to.
