+++ SuperNodes/src/models/LifecycleHooks.cs
@@ -1,0 +1,25 @@
+namespace SuperNodes.Models;
+
+using System.Collections.Immutable;
+
+/// <summary>
+/// Represents a lifecycle method handler provided by another source generator
+/// or a PowerUp's lifecycle method handler.
+/// </summary>
+public interface IGodotNodeLifecycleHook { }
+
+/// <summary>
+/// A string containing the name of the method that would be generated by
+/// another source generator or provided by a partial implementation.
+/// </summary>
+/// <param name="Name">Expected method name.</param>
+public record LifecycleMethodHook(string Name) : IGodotNodeLifecycleHook;
+
+/// <summary>
+/// Represents a lifecycle method handler provided by an applied PowerUp.
+/// </summary>
+/// <param name="FullName">Fully resolved name of the PowerUp.</param>
+/// <param name="TypeArguments">Type arguments provided to the PowerUp, if the
+/// PowerUp has generic parameters. Otherwise an empty array.</param>
+public record PowerUpHook(string FullName, ImmutableArray<string> TypeArguments)
+  : IGodotNodeLifecycleHook;
